refactor: ts
