+++ templates/ts/default-bot-message-extension/src/config.ts
