+++ templates/ts/default-bot-message-extension/src/index.ts
